Delete all of static files and re-import using build.py
This touches nearly all files because the permissions are reset to
non-executable.  This also deletes the non-minified version of
bootstrap.js and a few additional files that had found their way in.

The other big change is adding all of the .less files in for those who
want to use Less.js directly.